% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_f.R
\name{estimate_f}
\alias{estimate_f}
\title{Estimate F-statistic}
\usage{
estimate_f(samplesize_exposure, n_variants, rsq_exposure, lci_95 = FALSE)
}
\arguments{
\item{samplesize_exposure}{(numeric) Sample size of population used to define genetic instrument for the exposure of interest}

\item{n_variants}{(numeric) Number of genetic variants included in genetic instrument for the exposure of interest}

\item{rsq_exposure}{(numeric) \eqn{R^2} value (coefficient of determination) of genetic instrument for the exposure of interest; used to estimate}

\item{lci_95}{(logical; default = FALSE) If TRUE, the function will return the lower limit of the one-sided 95\% confidence interval of the F-statistic, which may represent a more conservative/less optimistic estimate}
}
\value{
Numeric vector containing the estimated F-statistic
}
\description{
Estimate F-statistic
}
\examples{
estimate_f(samplesize_exposure = 361194, n_variants = 196, rsq_exposure = 0.068, lci_95 = FALSE)

# return lower bound of one-sided 95\% confidence interval of F-statistic
estimate_f(samplesize_exposure = 361194, n_variants = 196, rsq_exposure = 0.068, lci_95 = TRUE)
}
