% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_overlap_bias.R
\name{estimate_overlap_bias}
\alias{estimate_overlap_bias}
\title{Estimate bias due to sample overlap}
\usage{
estimate_overlap_bias(
  samplesize_exposure,
  samplesize_outcome,
  n_variants,
  rsq_exposure,
  exp_f = NULL,
  lci_95 = FALSE,
  case_prop = 0,
  ols_bias,
  overlap_prop,
  var_x = 1,
  var_y = 1
)
}
\arguments{
\item{samplesize_exposure}{(numeric) Sample size of population used to define genetic instrument for the exposure of interest}

\item{samplesize_outcome}{(numeric) Sample size of population used for the outcome of interest}

\item{n_variants}{(numeric) Number of genetic variants included in genetic instrument for the exposure of interest}

\item{rsq_exposure}{(numeric) \eqn{R^2} value (coefficient of determination) of genetic instrument for the exposure of interest; used to estimate F-statistic}

\item{exp_f}{(numeric; optional) F-statistic for the genetic instrument (if provided, this value will be used, rather than an estimate based on the \eqn{R^2})}

\item{lci_95}{(logical; default = FALSE) If TRUE, the function will return estimates of bias and type 1 error based on the lower limit of the one-sided 95\% confidence interval of the F-statistic, which may represent a more conservative/less optimistic estimate of bias}

\item{case_prop}{(numeric; optional) Proportion of cases (eg. cases/total samplesize) if outcome is binary}

\item{ols_bias}{(numeric) Observational (biased) effect estimate (if known); otherwise, provide a hypothetical value}

\item{overlap_prop}{(numeric; range = 0 to 1) Proportion of overlapping samples between exposure and outcome studies (if known); otherwise, provide a hypothetical value}

\item{var_x}{(numeric) Variance in the exposure; default is 1 when the exposure is reported in standard deviation units}

\item{var_y}{(numeric) Variance in the exposure; default is 1 when the exposure is reported in standard deviation units}
}
\value{
A \link[tibble:tibble-package]{tibble} containing columns for the bias and type1_error
}
\description{
Estimate bias due to sample overlap
}
\examples{
# Binary outcome
estimate_overlap_bias(
  samplesize_exposure = 361194,
  samplesize_outcome = 1125328,
  case_prop = 0.035,
  rsq_exposure = 0.068,
  n_variants = 196,
  ols_bias = 0.2,
  overlap_prop = 0.3
)

# Continuous outcome
estimate_overlap_bias(
  samplesize_exposure = 361194,
  samplesize_outcome = 1125328,
  rsq_exposure = 0.068,
  n_variants = 196,
  ols_bias = 0.2,
  overlap_prop = 0.3
)
}
