% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrSampleOverlap-package.R
\docType{package}
\name{mrSampleOverlap-package}
\alias{mrSampleOverlap}
\alias{mrSampleOverlap-package}
\title{mrSampleOverlap: Estimate Bias Due To Sample Overlap In Mendelian Randomization Studies}
\description{
A function to estimate bias due to sample overlap in Mendelian Randomization studies.
}
\author{
\strong{Maintainer}: Levin Michael \email{michael.levin@pennmedicine.upenn.edu} (\href{https://orcid.org/0000-0002-9937-9932}{ORCID})

}
\keyword{internal}
